+++ packages/core/expressions/src/external-links.ts
@@ -1,3 +1,3 @@
 export default {
-  expressionsLanguageDoc: 'https://docs.konghq.com/gateway/latest/reference/expressions-language/language-references/',
+  expressionsLanguageDoc: 'https://developer.konghq.com/gateway/routing/expressions/',
 }