+++ packages/entities/entities-plugins/sandbox/index.ts
@@ -83,6 +83,11 @@ const init = async () => {
         name: 'configure-dynamic-ordering',
         component: () => import('./pages/FallbackPage.vue'),
       },
+      {
+        path: '/free-form/mocking',
+        name: 'free-form',
+        component: () => import('./pages/FreeFormPage.vue'),
+      },
     ],
   })