+++ packages/entities/entities-plugins/src/components/free-form/Datakit/examples/index.ts
@@ -1,2 +1,3 @@
 export { default as authenticate } from './authenticate.yml?raw'
 export { default as combine } from './combine.yml?raw'
+export { default as manipulate } from './manipulate.yml?raw'