+++ packages/entities/entities-plugins/src/types/plugin.ts
@@ -1,4 +1,4 @@
-import type { PathToDotNotation } from '@kong-ui-public/i18n/dist/types/types'
+import type { PathToDotNotation } from '@kong-ui-public/i18n'
 import { PluginGroup, PluginScope } from '@kong-ui-public/entities-plugins-metadata'
 
 export const PluginGroupArray = [
@@ -117,7 +117,6 @@ export type PluginMetaData<I18nMessageSource = void> = {
   group: PluginGroup // Plugin categories meta.
   scope: PluginScope[] // The scope supported by the Plugin.
   isEnterprise: boolean // The value will be True if the Plugin is enterprise only.
-  imageName?: string // An optional tag to define plugin's icon image.
   useLegacyForm?: boolean // An optional field to use legacy form for the plugin. Default to false.
   fieldRules?: FieldRules
 }