+++ packages/analytics/analytics-config-store/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.9.20...@kong-ui-public/analytics-config-store@1.0.1) (2025-07-16)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [0.9.20](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.9.19...@kong-ui-public/analytics-config-store@0.9.20) (2025-07-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [0.9.19](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.9.18...@kong-ui-public/analytics-config-store@0.9.19) (2025-07-03)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
 ## [0.9.18](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.9.17...@kong-ui-public/analytics-config-store@0.9.18) (2025-06-26)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-config-store