+++ packages/core/documentation/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@0.8.33...@kong-ui-public/documentation@1.1.0) (2025-07-16)
+
+
+### Features
+
+* **documentation:** release version 1.0.0 ([#2223](https://github.com/Kong/public-ui-components/issues/2223)) ([ad7ba4d](https://github.com/Kong/public-ui-components/commit/ad7ba4d763b710ee762ee832a3581446661fbe03))
+
+
+
+
+
+## [0.8.33](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@0.8.32...@kong-ui-public/documentation@0.8.33) (2025-07-16)
+
+**Note:** Version bump only for package @kong-ui-public/documentation
+
+
+
+
+
+## [0.8.32](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@0.8.31...@kong-ui-public/documentation@0.8.32) (2025-07-14)
+
+
+### Bug Fixes
+
+* move `moduleResolution` to `bundler` ([#2213](https://github.com/Kong/public-ui-components/issues/2213)) ([a72a0ce](https://github.com/Kong/public-ui-components/commit/a72a0ce0f5d326382c7fc5b9db47288cfbaaeb5b))
+
+
+
+
+
 ## [0.8.31](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@0.8.30...@kong-ui-public/documentation@0.8.31) (2025-06-26)
 
 **Note:** Version bump only for package @kong-ui-public/documentation