+++ packages/core/forms/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.13.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.13.6...@kong-ui-public/forms@4.13.7) (2025-07-16)
+
+**Note:** Version bump only for package @kong-ui-public/forms
+
+
+
+
+
+## [4.13.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.13.5...@kong-ui-public/forms@4.13.6) (2025-07-14)
+
+
+### Bug Fixes
+
+* move `moduleResolution` to `bundler` ([#2213](https://github.com/Kong/public-ui-components/issues/2213)) ([a72a0ce](https://github.com/Kong/public-ui-components/commit/a72a0ce0f5d326382c7fc5b9db47288cfbaaeb5b))
+
+
+
+
+
 ## [4.13.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.13.4...@kong-ui-public/forms@4.13.5) (2025-06-26)
 
 **Note:** Version bump only for package @kong-ui-public/forms