+++ packages/entities/entities-keys/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.8.33](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.8.32...@kong-ui-public/entities-keys@3.8.33) (2025-07-16)
+
+**Note:** Version bump only for package @kong-ui-public/entities-keys
+
+
+
+
+
+## [3.8.32](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.8.31...@kong-ui-public/entities-keys@3.8.32) (2025-07-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-keys
+
+
+
+
+
 ## [3.8.31](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.8.30...@kong-ui-public/entities-keys@3.8.31) (2025-06-26)
 
 **Note:** Version bump only for package @kong-ui-public/entities-keys