+++ packages/portal/document-viewer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.74](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/document-viewer@2.1.73...@kong-ui-public/document-viewer@2.1.74) (2025-07-14)
+
+**Note:** Version bump only for package @kong-ui-public/document-viewer
+
+
+
+
+
 ## [2.1.73](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/document-viewer@2.1.72...@kong-ui-public/document-viewer@2.1.73) (2025-06-26)
 
 **Note:** Version bump only for package @kong-ui-public/document-viewer